// src/contracts/generated/BrowserVerificationKeySetV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BrowserVerificationKeyV1 } from "./BrowserVerificationKeyV1";
import type { CatalogId } from "./CatalogId";
import type { CatalogVersion } from "./CatalogVersion";
import type { Digest32 } from "./Digest32";
import type { UtcInstant } from "./UtcInstant";

export type BrowserVerificationKeySetV1 = { schemaVersion: number, realmId: CatalogId, catalogVersion: CatalogVersion, realmCatalogDigest: Digest32, keySetDigest: Digest32, keys: Array<BrowserVerificationKeyV1>, publishedAt: UtcInstant, expiresAt: UtcInstant, };