// src/contracts/generated/CatalogRootPinV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CatalogId } from "./CatalogId";
import type { EcP256PublicJwkV1 } from "./EcP256PublicJwkV1";
import type { UtcInstant } from "./UtcInstant";

export type CatalogRootPinV1 = { keyId: CatalogId, publicJwk: EcP256PublicJwkV1, notBefore: UtcInstant, issueUntil: UtcInstant, verifyUntil: UtcInstant, emergencyDenied: boolean, };