// src/contracts/generated/EcP256PublicJwkV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Digest32 } from "./Digest32";
import type { EcKeyTypeV1 } from "./EcKeyTypeV1";
import type { P256CurveV1 } from "./P256CurveV1";

export type EcP256PublicJwkV1 = { kty: EcKeyTypeV1, crv: P256CurveV1, x: Digest32, y: Digest32, };