// src/contracts/generated/SignedAuthProjectionV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AuthProjectionV1 } from "./AuthProjectionV1";
import type { CatalogSignatureV1 } from "./CatalogSignatureV1";

export type SignedAuthProjectionV1 = { schemaVersion: number, payload: AuthProjectionV1, proof: CatalogSignatureV1, };