// src/contracts/generated/SignedBrowserVerificationKeySetV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BrowserVerificationKeySetV1 } from "./BrowserVerificationKeySetV1";
import type { CatalogSignatureV1 } from "./CatalogSignatureV1";

export type SignedBrowserVerificationKeySetV1 = { schemaVersion: number, payload: BrowserVerificationKeySetV1, proof: CatalogSignatureV1, };